Added agenta's logo in the favicon and set a title
