+++ agenta-web/src/components/Logo/Logo.tsx
@@ -1,16 +1,10 @@
 import Image from "next/image"
 import {useMemo} from "react"
 import {useAppTheme} from "../Layout/ThemeContextProvider"
+import useLogo from "@/lib/hooks/useLogo"
 const Logo: React.FC = () => {
-    const {appTheme} = useAppTheme()
+    const logoSrc = useLogo()
 
-    const logoSrc = useMemo(
-        () =>
-            appTheme === "dark"
-                ? "/assets/dark-complete-transparent-CROPPED.png"
-                : "/assets/light-complete-transparent-CROPPED.png",
-        [appTheme],
-    )
     return <Image src={logoSrc} alt="Agenta Logo" width={129} height={42.5} />
 }
